# apps/backend/src/main/java/com/travelplanner/api/openapi/openapi.yaml
...
openapi: 3.0.3
info:
  title: Travel Planner API
  version: 1.0.0
  description: |
    Knowledge-grounded, chat-first trip planner.

    **Errors.** Every non-2xx response returns the `ApiErrorResponse` envelope
    `{ code, message, details }`. `code` is a registered `snake_case` identifier; the frontend
    maps it to `common.errors.<code>` and never displays `message` verbatim.

    **Request correlation.** Send `X-Request-Id` (a UUID) on every request. The server echoes
    it on every response — success and failure alike — and puts it in the log MDC. If the
    header is absent or malformed the server generates one.

    **Pagination.** List endpoints take `page`, `page_size` and `sort`, and return the item
    array alongside `page`, `page_size` and `total`.

    **Optimistic concurrency.** Mutations of versioned aggregates carry `expected_version` in
    the request body (ADR 008). A mismatch is `409 version_conflict`; a missing value is
    `400 validation_failed` and is never treated as a force-overwrite.
  license:
    name: UNLICENSED
servers:
  - url: /api/v1
    description: Same-origin path prefix. ADR 006 routes the browser through the Next.js proxy.
...
paths:
...
  /trips:
    get:
      tags: [trips]
      operationId: listTrips
      summary: The caller's trips (UC-C1-01)
      description: |
        Every trip owned by the caller, newest first.

        **Not paginated.** One person's trips are a list they scroll, not a corpus, and shipping
        `page`/`page_size`/`total` that no client would vary is three fields of ceremony. The key
        stays `trips` if that ever changes.

        Each item carries its `version`, because a client that lists trips is one rename away from
        needing it and ADR 008 §1 requires the read before a write to supply it.
      security:
        - cookieAuth: []
      responses:
        '200':
          description: The caller's trips, newest first. Empty for a new account.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags: [trips]
      operationId: createTrip
      summary: Create a trip (UC-C1-01)
      description: |
        Creates a `DRAFT` trip and its empty brief in one transaction, in the documented lock order
        (`trip` then `trip_brief`). A trip whose brief row was missing would make every later read
        a "does it exist yet?" branch, and an empty brief is precisely what `DRAFT` means.

        The owner comes from the session cookie and is never in the body: an actor a request can
        assert is not an actor anything may be attributed to.

        No `expected_version` — ADR 008 versions edits to existing state, and there is none here
        for a concurrent writer to have replaced.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTripRequest'
      responses:
        '201':
          description: |
            The new trip, including the `version` its first brief save will have to echo — so the
            client does not need a second request to obtain it.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trip'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /trips/{tripId}:
    get:
      tags: [trips]
      operationId: getTrip
      summary: One trip
      description: |
        A trip owned by somebody else answers `404 not_found`, identically to one that does not
        exist. `TripRepositoryPort` publishes no unscoped lookup, so the two cases are
        indistinguishable here by construction rather than by discipline — a `403` would confirm
        that another user's trip exists to anyone who asked.
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/TripIdParam'
      responses:
        '200':
          description: The requested trip.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trip'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      tags: [trips]
      operationId: renameTrip
      summary: Rename a trip
      description: |
        The name is the only field a client may write on the trip row.

        **There is no `status` in the body, deliberately.** `trip.status` is derived from brief
        completeness by `PUT .../brief`, and moved to `ARCHIVED` by the archive action. A client
        able to set it directly could declare `BRIEF_COMPLETE` over an empty brief and walk past
        the gate PLAN §3.1 puts in front of C2.

        An archived trip is read-only for every actor, the agent included, so a rename on one is
        `400 validation_failed`.
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/TripIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenameTripRequest'
      responses:
        '200':
          $ref: '#/components/responses/TripMutated'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/VersionConflict'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags: [trips]
      operationId: deleteTrip
      summary: Delete a trip
      description: |
        Removes the trip and, by `ON DELETE CASCADE`, its brief.

        **No `expected_version`, and that is a decision rather than an omission.** ADR 008 versions
        writes that are *based on* prior content, so the loser of a race can re-read and re-apply
        its edit. A delete is based on nothing and has nothing to re-apply: the caller's intent is
        "this trip should not exist", and that does not become wrong because the agent changed a
        field a moment earlier. Carrying the version would also mean a request body on `DELETE`,
        which intermediaries are entitled to drop — a concurrency control that silently stops
        working behind some proxies is worse than none.

        `404` for a trip that is not the caller's, so a delete cannot be used to probe for one.
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/TripIdParam'
      responses:
        '204':
          description: The trip and its brief are gone. Nothing in the body — no resource is left.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /trips/{tripId}/actions/archive:
    post:
      tags: [trips]
      operationId: archiveTrip
      summary: Archive a trip
      description: |
        Moves the trip to `ARCHIVED`, after which it is view-only for every actor, the agent
        included (PLAN §3.1).

        A typed action rather than a `status` field on `PUT` (ADR 008 §3). Archiving is not a field
        edit, and giving it the same request shape as a rename is how one gets sent by accident.

        Not idempotent: archiving an already-archived trip is `400 validation_failed`, because the
        second call is a client acting on a stale view and ADR 008's whole position is that such a
        client should be told rather than quietly agreed with.

        It still carries `expected_version` — making a trip read-only while the agent is mid-edit
        loses that edit like any other write does.
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/TripIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArchiveTripRequest'
      responses:
        '200':
          $ref: '#/components/responses/TripMutated'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/VersionConflict'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: tp_session
      description: |
        The session cookie from ADR 002 — `httpOnly`, `SameSite=Lax`, and `Secure` in production.

        Documented as `apiKey in cookie` because OpenAPI 3.0 has no vocabulary for "the browser
        sends this automatically". No client ever sets this header or reads this value: it is
        `httpOnly`, and `lib/api/client.ts` relies on `credentials: 'include'` instead (ADR 006).
  headers:
...
    XRequestId:
      description: |
        Correlation id echoed from the request, or generated when the caller sent none.
        Present on every response, including errors — it is how a user-reported failure is
        found in the logs.
      required: true
      schema:
        type: string
        format: uuid
  parameters:
...
    TripIdParam:
      name: tripId
      in: path
      required: true
      description: |
        The trip's surrogate key. Ownership is *not* carried here — it comes from the session on
        every request, and an id that is not the caller's is answered `404 not_found` rather than
        `403`, so this parameter cannot be used to probe for another user's trips.
      schema:
        type: string
        format: uuid
...
  schemas:
...
    ApiErrorResponse:
      $ref: './errors.yaml#/components/schemas/ApiErrorResponse'
...
    TripStatus:
      type: string
      description: |
        Wizard progress for a trip — the exact vocabulary of the "Trip status" table in
        `USE-CASES.md`, in the order that table lists it (PLAN §3.1).

        **Read-only on the wire.** No request body anywhere on this API accepts a status. It is
        derived from brief completeness, and moved to `ARCHIVED` by the archive action; a client
        able to set it could declare `BRIEF_COMPLETE` over an empty brief and step past the gate in
        front of C2.

        C1 owns `DRAFT`, `CLARIFICATION_NEEDED`, and `BRIEF_COMPLETE`. The rest are published here
        because a client has to be able to render a trip in any of them, not because C1 can set
        them.
      enum:
        - DRAFT
        - BRIEF_COMPLETE
        - CLARIFICATION_NEEDED
        - RESEARCH_QUEUED
        - RESEARCH_RUNNING
        - RESEARCH_READY
        - DESTINATION_SELECTED
        - ITINERARY_READY
        - BOOKING_IN_PROGRESS
        - BOOKED
        - ARCHIVED
      example: DRAFT
    Trip:
      type: object
      description: |
        One trip, as every trip endpoint returns it.

        No `user_id`: ownership comes from the session on every request, and echoing the owner back
        would publish an identifier that no endpoint on this API accepts as input.
      required: [trip_id, name, status, version, created_at, updated_at]
      properties:
        trip_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 120
          example: Japan in spring
        status:
          $ref: '#/components/schemas/TripStatus'
        selected_recommendation_id:
          type: string
          format: uuid
          nullable: true
          description: The chosen recommendation. Absent until C2 completes (UC-C2-06).
        version:
          type: integer
          format: int32
          minimum: 0
          description: |
            The optimistic lock (ADR 008 §1). Published on every read because a client that cannot
            see the version cannot send `expected_version`, and would have no option but to force
            overwrite.
          example: 3
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TripList:
      type: object
      description: |
        The body of `GET /trips`.

        An object rather than a bare array: a top-level array cannot gain a field, and this one
        will want page metadata when a single user's trip list stops being a screenful. Wrapping
        now costs one key; wrapping later breaks every generated client.
      required: [trips]
      properties:
        trips:
          type: array
          description: Newest first. Empty for a new account, which is an answer rather than an error.
          items:
            $ref: '#/components/schemas/Trip'
    CreateTripRequest:
      type: object
      description: |
        The name is the only field. Everything else about a new trip is the server's: the owner
        from the session, `DRAFT` as the status, and an empty brief written in the same
        transaction.
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
          example: Japan in spring
    RenameTripRequest:
      allOf:
        - $ref: '#/components/schemas/ExpectedVersion'
        - type: object
          description: |
            No `status`. See the `TripStatus` note on why the field is not writable anywhere.
          required: [name]
          properties:
            name:
              type: string
              minLength: 1
              maxLength: 120
              example: Kyoto and Osaka
    ArchiveTripRequest:
      allOf:
        - $ref: '#/components/schemas/ExpectedVersion'
        - type: object
          description: |
            A body with one field rather than an empty `POST`: archiving is a mutation of a
            versioned aggregate like any other, and making a trip read-only while the agent is
            mid-edit loses that edit.
...
    ExpectedVersion:
      type: object
      description: |
        Optimistic-concurrency fragment (ADR 008 §2). Every mutation of a versioned aggregate
        composes this into its request body:

        ```yaml
        UpdateTripBriefRequest:
          allOf:
            - $ref: '#/components/schemas/ExpectedVersion'
            - type: object
              properties: { destinations: { ... } }
        ```

        The version travels in the body rather than in `ETag`/`If-Match` precisely because a
        body field is expressible in OpenAPI and flows through codegen into the typed client;
        header-based concurrency would need hand-plumbed handling in the generated-client
        layer this project forbids editing.

        Omitting `expected_version` is `400 validation_failed` — **never** a force-overwrite.
        A mismatch is `409 version_conflict` carrying `details.current_version`.
      required: [expected_version]
      properties:
        expected_version:
          type: integer
          format: int32
          minimum: 0
          description: The version the client based its edit on.
          example: 7
  responses:
    ValidationFailed:
      description: |
        The request failed schema or constraint validation. `details.fields` maps field name to
        an English message (`ValidationFailedDetails`).
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            code: validation_failed
            message: The request is not valid.
            details:
              fields:
                expected_version: must not be null
    Unauthorized:
      description: No valid session. The caller must sign in.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            code: unauthorized
            message: Authentication is required.
            details: {}
...
    NotFound:
      description: |
        No resource at this path, or none owned by the caller. Ownership failures return 404
        rather than 403 so the API does not confirm that another user's resource exists.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            code: not_found
            message: The requested resource does not exist.
            details: {}
...
    TripMutated:
      description: |
        The **new full resource** with its incremented `version` (ADR 008 §2). Returning the whole
        trip rather than an acknowledgement is what makes the conflict rules usable: the client
        holds exactly what the server holds, including the version its next write must echo.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Trip'
...
    VersionConflict:
      description: |
        The `expected_version` in the request body does not match the stored version
        (ADR 008). `details.current_version` carries the server's version so the client can
        re-read, re-apply its edit, and retry. There is no implicit winner.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            code: version_conflict
            message: The resource was modified by someone else.
            details:
              current_version: 9
...
    InternalError:
      description: Unhandled server fault. `details` is always empty — diagnostics stay in the logs.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            code: internal_error
            message: An unexpected error occurred.
            details: {}